refactor(Viewer): Move providers in a specific folder
